+++ src/modules-arch/schemes/reit.js
@@ -37,6 +37,7 @@ import { WithdrawalFiatCardModule } from '@modules/withdrawal-fiat-card/module'
 import { WithdrawalFiatBankModule } from '@modules/withdrawal-fiat-bank/module'
 import { RedeemFormModule } from '@modules/redeem-form/module'
 import { CoinpaymentsDepositModule } from '@/vue/modules/coinpayments-deposit/module'
+import { WithdrawalDrawerPseudoModule } from '@/modules-arch/pseudo-modules/withdrawal-drawer-pseudo-module'
 
 export default {
   importEnLocaleFile () {
@@ -119,6 +120,7 @@ export default {
                   new WithdrawalFiatBankModule(),
                 ],
               }),
+              new WithdrawalDrawerPseudoModule(),
               new CoinpaymentsDepositModule(),
               new DepositFiatModule({
                 submodules: [